wip: redesign meta reports
