+++ app/public/src/pages/component/meta-report/meta-report.tsx
@@ -1,295 +1,44 @@
-import React, { useRef, useState } from "react"
-import TeamComp from "./team-comp"
+import React, { useRef } from "react"
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs"
 
-import ItemStatistic from "./item-statistic"
-import { IMeta } from "../../../../../models/mongo-models/meta"
-import { IItemsStatistic } from "../../../../../models/mongo-models/items-statistic"
-import PokemonStatistic from "./pokemon-statistic"
-import { IPokemonsStatistic } from "../../../../../models/mongo-models/pokemons-statistic"
 import { useTranslation } from "react-i18next"
 import { requestMeta } from "../../../stores/NetworkStore"
-import { useAppDispatch, useAppSelector } from "../../../hooks"
-import { Synergy } from "../../../../../types/enum/Synergy"
-import { Rarity } from "../../../../../types/enum/Game"
-
-const optStyle = {
-  color: "black"
-}
+import { useAppDispatch } from "../../../hooks"
+import { CompositionReport } from "./composition-report"
+import { ItemReport } from "./item-report"
+import { PokemonReport } from "./pokemon-report"
+import "./meta-report.css"
 
 export default function MetaReport() {
   const { t } = useTranslation()
   const dispatch = useAppDispatch()
 
-  const [rankingBy, setRanking] = useState<string>("count")
-  const [itemRankingBy, setItemRanking] = useState<string>("count")
-  const [pokemonRankingBy, setPokemonRanking] = useState<string>("count")
-  const [synergy, setSynergy] = useState<Synergy | "all">("all")
-  const [rarity, setRarity] = useState<Rarity | "all">("all")
   const hasLoadedMeta = useRef<boolean>(false)
-
-  const meta = useAppSelector((state) => state.lobby.meta)
-  const metaItems = useAppSelector((state) => state.lobby.metaItems)
-  const metaPokemons = useAppSelector((state) => state.lobby.metaPokemons)
 
   if (hasLoadedMeta.current === false) {
     dispatch(requestMeta())
     hasLoadedMeta.current = true
   }
 
-  let sortedMeta = new Array<IMeta>()
-  let sortedMetaItems = new Array<IItemsStatistic>()
-  let sortedMetaPokemons = new Array<IPokemonsStatistic>()
-  if (rankingBy == "count" || rankingBy == "winrate") {
-    sortedMeta = [...meta].sort((a, b) => {
-      return b[rankingBy] - a[rankingBy]
-    })
-  } else {
-    sortedMeta = [...meta].sort((a, b) => {
-      return a[rankingBy] - b[rankingBy]
-    })
-  }
-  if (itemRankingBy == "count") {
-    sortedMetaItems = [...metaItems].sort((a, b) => {
-      return b[itemRankingBy] - a[itemRankingBy]
-    })
-  } else {
-    sortedMetaItems = [...metaItems].sort((a, b) => {
-      return a[itemRankingBy] - b[itemRankingBy]
-    })
-  }
-  if (pokemonRankingBy == "count") {
-    sortedMetaPokemons = [...metaPokemons].sort((a, b) => {
-      return b[pokemonRankingBy] - a[pokemonRankingBy]
-    })
-  } else {
-    sortedMetaPokemons = [...metaPokemons].sort((a, b) => {
-      return a[pokemonRankingBy] - b[pokemonRankingBy]
-    })
-  }
   return (
-    <div>
-      <div
-        className="nes-container"
-        style={{
-          margin: "10px",
-          height: "90vh",
-          color: "white",
-          width: "80vw"
-        }}
-      >
-        <Tabs>
-          <TabList>
-            <Tab key="team-comps">{t("meta_report")}</Tab>
-            <Tab key="items">{t("item_report")}</Tab>
-            <Tab key="pokemons">{t("pokemon_report")}</Tab>
-          </TabList>
+    <div id="meta-report" className="nes-container">
+      <Tabs>
+        <TabList>
+          <Tab key="team-comps">{t("meta_report")}</Tab>
+          <Tab key="items">{t("item_report")}</Tab>
+          <Tab key="pokemons">{t("pokemon_report")}</Tab>
+        </TabList>
 
-          <TabPanel key="team-comps-panel">
-            <div
-              style={{
-                display: "flex",
-                justifyContent: "space-between",
-                alignItems: "center",
-                paddingRight: "27px",
-                paddingLeft: "15px"
-              }}
-            >
-              <h3>{t("best_team_compositions")}</h3>
-              <div
-                style={{
-                  display: "flex",
-                  width: "23%",
-                  alignItems: "center",
-                  justifyContent: "space-around",
-                  backgroundColor: "rgb(84, 89, 107)"
-                }}
-                className="my-select"
-              >
-                <p style={{ margin: "0px" }}>{t("rank")}</p>
-                <select
-                  value={rankingBy}
-                  onChange={(e) => {
-                    setRanking(e.target.value)
-                  }}
-                  style={{ background: "none", border: "none", color: "white" }}
-                >
-                  <option style={optStyle} value="count">
-                    {t("by_poularity")}
-                  </option>
-                  <option style={optStyle} value="mean_rank">
-                    {t("by_average_place")}
-                  </option>
-                  <option style={optStyle} value="winrate">
-                    {t("by_winrate")}
-                  </option>
-                </select>
-              </div>
-            </div>
-
-            <div style={{ height: "70vh", overflowY: "scroll" }}>
-              {sortedMeta.length === 0 && <p>No data available</p>}
-              {sortedMeta.map((team) => {
-                return <TeamComp team={team} key={team.cluster_id} />
-              })}
-            </div>
-          </TabPanel>
-          <TabPanel>
-            <div
-              style={{
-                display: "flex",
-                justifyContent: "space-between",
-                alignItems: "center",
-                paddingRight: "27px",
-                paddingLeft: "15px"
-              }}
-            >
-              <h3>{t("best_items")}</h3>
-              <div
-                style={{
-                  display: "flex",
-                  width: "23%",
-                  alignItems: "center",
-                  justifyContent: "space-around",
-                  backgroundColor: "rgb(84, 89, 107)"
-                }}
-                className="my-select"
-              >
-                <p style={{ margin: "0px" }}>{t("rank")}</p>
-                <select
-                  value={itemRankingBy}
-                  onChange={(e) => {
-                    setItemRanking(e.target.value)
-                  }}
-                  style={{ background: "none", border: "none", color: "white" }}
-                >
-                  <option style={optStyle} value="count">
-                    {t("by_popularity")}
-                  </option>
-                  <option style={optStyle} value="rank">
-                    {t("by_average_place")}
-                  </option>
-                </select>
-              </div>
-            </div>
-            <div style={{ height: "70vh", overflowY: "scroll" }}>
-              {sortedMetaItems.length === 0 && <p>No data available</p>}
-              {sortedMetaItems.map((item) => {
-                return <ItemStatistic item={item} key={item.name} />
-              })}
-            </div>
-          </TabPanel>
-          <TabPanel>
-            <div
-              style={{
-                display: "flex",
-                justifyContent: "space-between",
-                alignItems: "center",
-                paddingRight: "27px",
-                paddingLeft: "15px"
-              }}
-            >
-              <h3>{t("best_pokemons")}</h3>
-              <div
-                style={{
-                  display: "flex",
-                  width: "23%",
-                  alignItems: "center",
-                  justifyContent: "space-around",
-                  backgroundColor: "rgb(84, 89, 107)"
-                }}
-                className="my-select"
-              >
-                <p style={{ margin: "0px" }}>{t("rank")}</p>
-                <select
-                  value={pokemonRankingBy}
-                  onChange={(e) => {
-                    setPokemonRanking(e.target.value)
-                  }}
-                  style={{ background: "none", border: "none", color: "white" }}
-                >
-                  <option style={optStyle} value="count">
-                    {t("by_popularity")}
-                  </option>
-                  <option style={optStyle} value="rank">
-                    {t("by_average_place")}
-                  </option>
-                </select>
-              </div>
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-around",
-                  backgroundColor: "rgb(84, 89, 107)"
-                }}
-                className="my-select"
-              >
-                <p style={{ margin: "0px" }}>{t("synergies")}</p>
-                <select
-                  value={synergy}
-                  onChange={(e) => {
-                    setSynergy(e.target.value as any)
-                  }}
-                  style={{
-                    background: "none",
-                    border: "none",
-                    color: "white"
-                  }}
-                >
-                  <option style={optStyle} value={"all"}>
-                    All
-                  </option>
-                  {Object.keys(Synergy).map((s) => (
-                    <option style={optStyle} value={s}>
-                      {t(`synergy.${s}`)}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-around",
-                  backgroundColor: "rgb(84, 89, 107)"
-                }}
-                className="my-select"
-              >
-                <p style={{ margin: "0px" }}>{t("rarity_label")}</p>
-                <select
-                  value={rarity}
-                  onChange={(e) => {
-                    setRarity(e.target.value as any)
-                  }}
-                  style={{
-                    background: "none",
-                    border: "none",
-                    color: "white"
-                  }}
-                >
-                  <option style={optStyle} value={"all"}>
-                    All
-                  </option>
-                  {Object.keys(Rarity).map((s) => (
-                    <option style={optStyle} value={s}>
-                      {t(`rarity.${s}`)}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            </div>
-            {
-              <PokemonStatistic
-                pokemons={sortedMetaPokemons}
-                rankingBy={pokemonRankingBy}
-                synergy={synergy}
-                rarity={rarity}
-              />
-            }
-          </TabPanel>
-        </Tabs>
-      </div>
+        <TabPanel key="team-comps-panel">
+          <CompositionReport />
+        </TabPanel>
+        <TabPanel>
+          <ItemReport />
+        </TabPanel>
+        <TabPanel>
+          <PokemonReport />
+        </TabPanel>
+      </Tabs>
     </div>
   )
 }
